requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
tiktoken>=0.5.0
scikit-learn>=1.3.0
openai>=1.6.0
pyyaml>=6.0
pytest>=7.4.0
httpx>=0.25.0
